Optimise generated module code to reduce performance impact
`handle_event/4` function is *very* time sensitive function, especially
in high throughput systems where it is called a lot, so every single bit
of performance matters.

Here we are using more direct approach to iterate over list with
`:lists.foreach/2` which is BIF and avoids dynamic dispatch, as we know
that `metrics` is always list.

It inlines function calls to reduce overhead and well as optimise
fetching tags values from metadata. It slightly changes behaviour there
as non-existent keys will be non-existent in metadata instead of
containing empty strings.